+++ app/instructors/loading.tsx
@@ -1,0 +1,9 @@
+import { SkeletonTable } from '@/components/skeletons';
+
+export default function Loading() {
+  return (
+    <div className="mx-auto max-w-7xl px-4 py-4 sm:px-6 md:py-8 lg:px-8">
+      <SkeletonTable rows={6} columns={5} />
+    </div>
+  );
+}
