+++ features/instructors/index.ts
@@ -1,0 +1,2 @@
+// クエリフック
+export * from './api/queries';
